+++ src/components/date-picker/date-picker.vue
@@ -4,7 +4,7 @@
       v-model:shown="datePopperState"
       aria-id="date-picker-wrapper"
       distance="4"
-      placement="bottom"
+      :placement="props.placement"
       :triggers="[]"
       :popper-hide-triggers="[]"
       :auto-hide="false"
@@ -18,7 +18,7 @@
         width: props.width,
       }"
     >
-      <div :id="props.id" class="spr-grid spr-gap-2">
+      <div :id="props.id" class="spr-grid spr-gap-size-spacing-4xs">
         <label v-if="props.label" :for="props.id" :class="datePickerClasses.labelClasses">
           {{ props.label }}
         </label>
@@ -33,10 +33,9 @@
               maxlength="3"
               :disabled="props.disabled"
               :readonly="props.readonly"
-              @input="handleMonthInput(monthInput, $event as KeyboardEvent)"
-              @keyup="handleMonthInput(monthInput, $event as KeyboardEvent)"
-              @keydown="handleMonthInput(monthInput, $event as KeyboardEvent)"
-              @blur="handleMonthInput(monthInput, $event as FocusEvent)"
+              @input="handleMonthInput"
+              @keyup="handleMonthInput"
+              @keydown="handleBackspace('month', $event)"
             />
             <span class="spr-text-color-strong spr-font-size-200 spr-text-color-weak">/</span>
             <input
@@ -48,8 +47,9 @@
               maxlength="2"
               :disabled="props.disabled"
               :readonly="props.readonly"
-              @input="handleDateInput(dateInput, null, null, $event as KeyboardEvent)"
-              @keyup="handleDateInput(dateInput, null, null, $event as KeyboardEvent)"
+              @input="handleDateInput"
+              @keyup="handleDateInput"
+              @keydown="handleBackspace('date', $event)"
             />
             <span class="spr-text-color-strong spr-font-size-200 spr-text-color-weak">/</span>
             <input
@@ -61,12 +61,13 @@
               maxlength="4"
               :disabled="props.disabled"
               :readonly="props.readonly"
-              @input="handleYearInput(yearInput, $event as KeyboardEvent)"
-              @keyup="handleYearInput(yearInput, $event as KeyboardEvent)"
+              @input="handleYearInput"
+              @keyup="handleYearInput"
+              @keydown="handleBackspace('year', $event)"
             />
           </div>
           <div class="spr-flex spr-items-center spr-justify-center">
-            <Icon class="spr-h-5 spr-w-5 spr-text-mushroom-300" icon="ph:calendar-blank" />
+            <Icon class="spr-text-color-supporting spr-h-4 spr-w-4" icon="ph:calendar-blank" />
           </div>
         </div>
       </div>
@@ -76,7 +77,7 @@
           <div
             :class="[
               'spr-flex spr-justify-between spr-gap-2 spr-px-4 spr-py-3',
-              'spr-border spr-border-solid spr-border-mushroom-200',
+              'spr-border spr-border-x-0 spr-border-b spr-border-t-0 spr-border-solid spr-border-mushroom-200',
             ]"
           >
             <!-- Tabs -->
@@ -144,7 +145,7 @@
           </div>
           <div class="spr-px-4 spr-pb-4 spr-pt-2">
             <!-- Calendar Tab  -->
-            <div v-if="currentTab === 'tab-calendar'" class="spr-grid spr-grid-cols-7 spr-gap-2">
+            <div v-if="currentTab === 'tab-calendar'" class="spr-grid spr-grid-cols-7">
               <div
                 v-for="(dayOfWeek, dayOfWeekIndex) in daysOfWeek"
                 :key="dayOfWeekIndex"
@@ -172,7 +173,7 @@
                       'spr-text-color-disabled': calendarTabIsInactiveMonthDates(day),
 
                       // Selected Date
-                      'spr-border-color-brand-base spr-background-color-single-active active:spr-background-color-brand-pressed spr-cursor-pointer spr-border spr-border-solid active:spr-scale-95':
+                      'spr-background-color-brand-base active:spr-background-color-brand-pressed spr-text-color-inverted-strong spr-cursor-pointer spr-text-white-50 active:spr-scale-95':
                         calendarTabIsSelectedDate(day),
 
                       // Unselected Date
@@ -248,22 +249,12 @@
         </div>
       </template>
     </Menu>
-    <template v-if="datePickerErrors.length > 0">
-      <div v-if="props.displayHelper" :class="datePickerClasses.datePickerInputHelperClasses">
-        <slot name="helperMessage">
-          <Icon icon="ph:warning-circle-fill" width="20px" height="20px" />
-          <span>{{ datePickerErrors[0].message }}</span>
-        </slot>
-      </div>
-    </template>
-    <template v-else>
-      <div v-if="props.displayHelper" :class="datePickerClasses.datePickerInputHelperClasses">
-        <slot name="helperMessage">
-          <Icon v-if="props.helperIcon" :icon="props.helperIcon" width="20px" height="20px" />
-          <span>{{ props.helperText }}</span>
-        </slot>
-      </div>
-    </template>
+    <div v-if="props.displayHelper" :class="datePickerClasses.datePickerInputHelperClasses">
+      <slot name="helperMessage">
+        <Icon v-if="props.helperIcon" :icon="props.helperIcon" width="20px" height="20px" />
+        <span>{{ props.helperText }}</span>
+      </slot>
+    </div>
   </div>
 </template>
 
@@ -295,7 +286,6 @@ const {
   monthsList,
   monthInput,
   yearInput,
-  datePickerErrors,
   calendarTabPageData,
   calendarTabIsMinMonth,
   calendarTabIsMaxMonth,
@@ -323,5 +313,6 @@ const {
   handleDateInput,
   handleYearInput,
   handleTabClick,
+  handleBackspace,
 } = useDatePicker(props, emit);
 </script>