+++ src/components/filter/filter.vue
@@ -6,32 +6,42 @@
     placement="bottom"
     :triggers="['click']"
     :popper-hide-triggers="[]"
-    :container="`#${props.id}`"
-    :auto-hide="false"
+    :container="`#${uniqueId}`"
     :style="{
       width: props.width,
+      position: 'relative',
     }"
+    :auto-hide="false"
   >
-    <span :id="props.id">
-      <spr-input
-        v-model="searchValue"
-        type="text"
-        :placeholder="placeholder"
-        :label="label"
-        :disabled="disabled"
-        :style="{
-          width: props.width,
-        }"
-      >
-        <template #icon>
-          <Icon icon="ph:magnifying-glass" />
-        </template>
-      </spr-input>
+    <span
+      :id="uniqueId"
+      :style="{
+        width: props.width,
+        position: 'relative',
+      }"
+    >
+      <slot>
+        <spr-input v-model="searchValue" type="text" :placeholder="placeholder" :label="label" :disabled="disabled">
+          <template #icon>
+            <Icon icon="ph:magnifying-glass" />
+          </template>
+        </spr-input>
+      </slot>
     </span>
 
     <template #popper>
       <div :class="filterClass.MenuOptionClasses">
         <div v-if="filterMenu.length > 0 && filterable" class="spr-flex spr-items-center spr-gap-2">
+          <div class="spr-space-x-size-spacing-3xs spr-space-y-size-spacing-3xs">
+            <template v-for="menu in filterMenu" :key="menu.field">
+              <spr-chips
+                v-if="mappedFilterMenuList[menu.field].count"
+                :label="mappedFilterMenuList[menu.field].columnName"
+                icon="ph:funnel-simple"
+              />
+            </template>
+          </div>
+
           <Menu
             v-model:shown="isAddFilterVisible"
             aria-id="filter-menu-wrapper"
@@ -106,7 +116,8 @@
                           />
                         </div>
                         <div
-                          v-if="!loading && getFiltereredMenuOption.length > 0"
+                          v-if="!props.loading && getFiltereredMenuOption.length > 0"
+                          ref="filterMenuOptionList"
                           class="spr-max-h-[264px] spr-space-y-size-spacing-6xs spr-overflow-auto spr-p-size-spacing-2xs"
                         >
                           <div
@@ -128,7 +139,7 @@
                         </div>
 
                         <div v-else :class="filterClass.LoadingStateClasses">
-                          <slot v-if="loading" name="loading"> loading... </slot>
+                          <slot v-if="props.loading" name="loading"> loading... </slot>
                           <slot v-else name="empty"> No Result Found! </slot>
                         </div>
 
@@ -151,16 +162,6 @@
               </div>
             </template>
           </Menu>
-
-          <div class="spr-space-x-size-spacing-3xs spr-space-y-size-spacing-3xs">
-            <template v-for="menu in filterMenu" :key="menu.field">
-              <spr-chips
-                v-if="mappedFilterMenuList[menu.field].count"
-                :label="mappedFilterMenuList[menu.field].columnName"
-                icon="ph:funnel-simple"
-              />
-            </template>
-          </div>
         </div>
 
         <spr-button variant="secondary" size="small" @click="selectAllOptions">Select All</spr-button>
@@ -168,6 +169,7 @@
 
       <div
         v-if="getFiltereredOption.length > 0 && !filling"
+        ref="filterOptionRef"
         class="spr-max-h-[264px] spr-space-y-size-spacing-6xs spr-overflow-auto spr-p-size-spacing-3xs"
       >
         <div
@@ -225,6 +227,9 @@ const {
   filterMenuSearchvalue,
   mappedFilterMenuList,
   filterClass,
+  uniqueId,
+  filterOptionRef,
+  filterMenuOptionList,
 
   selectAllOptions,
   getMappedFilterData,