+++ src/components/input/input-contact-number/use-input-contact-number.ts
@@ -1,45 +1,105 @@
-import { ref, SetupContext } from 'vue';
-import { CountryOption, type InputContactNumberEmitTypes, type InputContactNumberPropTypes } from './input-contact-number';
-import parsePhoneNumber, { CountryCode } from 'libphonenumber-js'
-import { useVModel } from '@vueuse/core'
+import { ref, toRefs, computed, ComputedRef, onMounted, SetupContext } from 'vue';
+import { useVModel } from '@vueuse/core';
+
+import classNames from 'classnames';
+
+import parsePhoneNumber, { CountryCode } from 'libphonenumber-js';
+
+import {
+  CountryOption,
+  type InputContactNumberEmitTypes,
+  type InputContactNumberPropTypes,
+} from './input-contact-number';
+
+interface InputContactNumberClasses {
+  countryCallingCodeClasses: string;
+}
 
 export const useInputContactNumber = (
   props: InputContactNumberPropTypes,
   emit: SetupContext<InputContactNumberEmitTypes>['emit'],
 ) => {
+  const { disabled } = toRefs(props);
+
+  const inputContactNumberClasses: ComputedRef<InputContactNumberClasses> = computed(() => {
+    const countryCallingCodeClasses = classNames(
+      'spr-font-weight-regular spr-font-size-200 spr-line-height-500 spr-letter-spacing-none spr-font-main',
+      'spr-flex spr-items-center spr-gap-size-spacing-5xs',
+      {
+        'spr-cursor-not-allowed': disabled.value,
+        'spr-cursor-pointer': !disabled.value,
+      },
+    );
+
+    return {
+      countryCallingCodeClasses,
+    };
+  });
+
   const formattedValue = useVModel(props, 'modelValue', emit);
 
-  // const formattedValue = ref('');
-  const selectedCountry = ref<string[]>(['PH']); // Default to Philippines
+  const selectedCountry = ref({
+    countryCode: ['PH'],
+    countryCallingCode: ['63'],
+  });
 
-  const handleSelectedCountries = (item: CountryOption) => {
-    // Set the selected country
-    selectedCountry.value[0] = item.value;
+  const popperState = ref(false);
 
-    if (!formattedValue.value) return;
-    let newPhoneNumber = formattedValue.value;
+  const handleContactNumberInput = (event: InputEvent) => {
+    const input = event.target as HTMLInputElement;
+    const value = input.value;
 
-    // Remove the country code if the value has formatted value
-    if(formattedValue.value !== '') { 
-      newPhoneNumber = newPhoneNumber.substring(formattedValue.value.indexOf(" ")); 
+    if (!/^[+]?\d*[\d\s-]*$/.test(value)) {
+      input.value = value.replace(/[^0-9+\s-]/g, '');
     }
-  
-    // Parse the phone number
-    const phoneNumber = parsePhoneNumber(newPhoneNumber, { defaultCountry: selectedCountry.value[0] as CountryCode });
-    // Format the phone number
-    if (phoneNumber) {
-      formattedValue.value = phoneNumber.formatInternational();
+
+    emit('getContactNumberErrors', []);
+
+    if (value.length > 0) {
+      handleContactNumberInputFormat();
     }
   };
-  
-  const handleOnBlur = () => {
+
+  const handleContactNumberInputFormat = () => {
+    formatContactNumber();
+  };
+
+  const handleSelectedCountries = (item: CountryOption) => {
+    selectedCountry.value = {
+      countryCode: [item.countryCode],
+      countryCallingCode: [item.value],
+    };
+
+    emit('getContactNumberErrors', []);
+
+    formatContactNumber();
+
+    emit('getSelectedCountryCallingCode', selectedCountry.value.countryCallingCode[0]);
+  };
+
+  const formatContactNumber = () => {
     if (!formattedValue.value) return;
-  
-    const phoneNumber = parsePhoneNumber(formattedValue.value, { defaultCountry: selectedCountry.value[0] as CountryCode, extract: false});
-    if (phoneNumber) {
-        formattedValue.value = phoneNumber.formatInternational();
+
+    const normalizedNumber = formattedValue.value.replace(/\D/g, '');
+
+    const phoneNumber = parsePhoneNumber(normalizedNumber, {
+      defaultCountry: selectedCountry.value.countryCode[0] as CountryCode,
+      extract: false,
+    });
+
+    if (phoneNumber && phoneNumber.isValid()) {
+      let formattedNumber = phoneNumber.formatInternational();
+
+      formattedNumber = formattedNumber.replace(`+${selectedCountry.value.countryCallingCode[0]} `, '');
+
+      formattedValue.value = formattedNumber;
     } else {
-      formattedValue.value = '';
+      emit('getContactNumberErrors', [
+        {
+          title: 'Invalid Contact Number',
+          message: 'Must be a valid contact number',
+        },
+      ]);
     }
   };
 
@@ -47,11 +107,23 @@ export const useInputContactNumber = (
     emit('update:modelValue', value);
   };
 
-  return {
-    formattedValue,
-    handleSelectedCountries,
-    selectedCountry,
-    handleOnBlur,
-    handleUpdateModelValue,
+  const handlePopperState = (state: boolean) => {
+    popperState.value = state;
   };
-}; 
+
+  onMounted(() => {
+    emit('getSelectedCountryCallingCode', selectedCountry.value.countryCallingCode[0]);
+  });
+
+  return {
+    inputContactNumberClasses,
+    formattedValue,
+    selectedCountry,
+    popperState,
+    handleContactNumberInput,
+    handleSelectedCountries,
+    formatContactNumber,
+    handleUpdateModelValue,
+    handlePopperState,
+  };
+};