+++ src/components/input/input-search/input-search.vue
@@ -1,12 +1,13 @@
 <template>
-  <SprInput v-bind="$attrs">
+  <spr-input v-bind="$attrs">
     <template #icon>
-      <Icon icon="ph:magnifying-glass"/>
+      <Icon icon="ph:magnifying-glass" />
     </template>
-  </SprInput>
+  </spr-input>
 </template>
 
 <script setup lang="ts">
+import { Icon } from '@iconify/vue';
+
 import SprInput from '@/components/input/input.vue';
-import { Icon } from "@iconify/vue";
 </script>