+++ src/components/table/table.ts
@@ -1,5 +1,6 @@
 import type { PropType, ExtractPropTypes } from 'vue';
-
+import type { ChipTitle } from '@/components/table/table-chips-title/table-chips-title';
+import type { LozengeTitle } from '@/components/table/table-lozenge-title/table-lozenge-title';
 export const definePropType = <T>(val: unknown): PropType<T> => val as PropType<T>;
 
 interface Header {
@@ -7,16 +8,21 @@ interface Header {
   name: string;
   sort: boolean;
   hasAvatar: boolean;
+  hasIcon: boolean;
   hasSubtext: boolean;
+  hasLozengeTitle: boolean;
+  hasChipTitle: boolean;
   badgeText: string;
   badgeVariant: string;
+  avatarVariant: string;
 }
 
 interface TableData {
   [key: string]: {
-    title: string;
+    title: string | LozengeTitle | ChipTitle | LozengeTitle[] | ChipTitle[];
     subtext?: string;
     image?: string;
+    icon?: string;
   };
 }
 
@@ -64,7 +70,6 @@ export const tablePropTypes = {
     type: Array as PropType<Header[]>,
     required: true,
   },
-
   emptyState: {
     type: Object as PropType<EmptyState>,
     default: () => ({
@@ -113,8 +118,16 @@ export const tablePropTypes = {
 
   isRowClickable: {
     type: Boolean as PropType<boolean>,
-    default: false
-  }
+    default: false,
+  },
+  fullHeight: {
+    type: Boolean as PropType<boolean>,
+    default: true,
+  },
+  removeHeaderOnEmpty: {
+    type: Boolean as PropType<boolean>,
+    default: false,
+  },
 };
 
 export const tableEmitTypes = {
@@ -122,7 +135,9 @@ export const tableEmitTypes = {
   onSort: (value: SortEvent): value is SortEvent =>
     typeof value.field === 'string' && TABLE_SORT.includes(value.sortOrder),
   onRowClick: (rowData: TableData, rowKey: number): rowData is TableData =>
-    typeof rowData === 'object' && typeof rowKey === 'number'
+    typeof rowData === 'object' && typeof rowKey === 'number',
+  onHover: (value: { active: boolean; data: TableData }): value is { active: boolean; data: TableData } =>
+    typeof value.active === 'boolean' && typeof value.data === 'object',
 };
 
 export type TablePropTypes = ExtractPropTypes<typeof tablePropTypes>;